requirements.txt
transformers==4.48.0
torch
pillow
pdf2image
streamlit 